feat(primitives/sandbox): move tests out of imp.rs
