+++ primitives/sandbox/src/imp.rs
@@ -18,13 +18,13 @@
 use sp_std::collections::btree_map::BTreeMap;
 use sp_std::fmt;
 
+use super::{Error, HostError, HostFuncType, ReturnValue, Value};
+use wasmi::memory_units::Pages;
 use wasmi::{
 	Externals, FuncInstance, FuncRef, GlobalDescriptor, GlobalRef, ImportResolver,
-	MemoryDescriptor, MemoryInstance, MemoryRef, Module, ModuleInstance, ModuleRef,
-	RuntimeArgs, RuntimeValue, Signature, TableDescriptor, TableRef, Trap, TrapKind
+	MemoryDescriptor, MemoryInstance, MemoryRef, Module, ModuleInstance, ModuleRef, RuntimeArgs,
+	RuntimeValue, Signature, TableDescriptor, TableRef, Trap, TrapKind,
 };
-use wasmi::memory_units::Pages;
-use super::{Error, Value, ReturnValue, HostFuncType, HostError};
 
 #[derive(Clone)]
 pub struct Memory {
@@ -37,17 +37,22 @@ impl Memory {
 			memref: MemoryInstance::alloc(
 				Pages(initial as usize),
 				maximum.map(|m| Pages(m as usize)),
-			).map_err(|_| Error::Module)?,
+			)
+			.map_err(|_| Error::Module)?,
 		})
 	}
 
 	pub fn get(&self, ptr: u32, buf: &mut [u8]) -> Result<(), Error> {
-		self.memref.get_into(ptr, buf).map_err(|_| Error::OutOfBounds)?;
+		self.memref
+			.get_into(ptr, buf)
+			.map_err(|_| Error::OutOfBounds)?;
 		Ok(())
 	}
 
 	pub fn set(&self, ptr: u32, value: &[u8]) -> Result<(), Error> {
-		self.memref.set(ptr, value).map_err(|_| Error::OutOfBounds)?;
+		self.memref
+			.set(ptr, value)
+			.map_err(|_| Error::OutOfBounds)?;
 		Ok(())
 	}
 }
@@ -68,9 +73,7 @@ impl<T> Clone for DefinedHostFunctions<T> {
 
 impl<T> DefinedHostFunctions<T> {
 	fn new() -> DefinedHostFunctions<T> {
-		DefinedHostFunctions {
-			funcs: Vec::new(),
-		}
+		DefinedHostFunctions { funcs: Vec::new() }
 	}
 
 	fn define(&mut self, f: HostFuncType<T>) -> HostFuncIndex {
@@ -102,7 +105,8 @@ impl<'a, T> Externals for GuestExternals<'a, T> {
 		index: usize,
 		args: RuntimeArgs,
 	) -> Result<Option<RuntimeValue>, Trap> {
-		let args = args.as_ref()
+		let args = args
+			.as_ref()
 			.iter()
 			.cloned()
 			.map(Into::into)
@@ -244,9 +248,8 @@ impl<T> Instance<T> {
 		state: &mut T,
 	) -> Result<Instance<T>, Error> {
 		let module = Module::from_buffer(code).map_err(|_| Error::Module)?;
-		let not_started_instance = ModuleInstance::new(&module, env_def_builder)
-			.map_err(|_| Error::Module)?;
-
+		let not_started_instance =
+			ModuleInstance::new(&module, env_def_builder).map_err(|_| Error::Module)?;
 
 		let defined_host_functions = env_def_builder.defined_host_functions.clone();
 		let instance = {
@@ -254,7 +257,8 @@ impl<T> Instance<T> {
 				state,
 				defined_host_functions: &defined_host_functions,
 			};
-			let instance = not_started_instance.run_start(&mut externals)
+			let instance = not_started_instance
+				.run_start(&mut externals)
 				.map_err(|_| Error::Execution)?;
 			instance
 		};
@@ -278,8 +282,7 @@ impl<T> Instance<T> {
 			state,
 			defined_host_functions: &self.defined_host_functions,
 		};
-		let result = self.instance
-			.invoke_export(&name, &args, &mut externals);
+		let result = self.instance.invoke_export(&name, &args, &mut externals);
 
 		match result {
 			Ok(None) => Ok(ReturnValue::Unit),
@@ -289,186 +292,8 @@ impl<T> Instance<T> {
 	}
 
 	pub fn get_global_val(&self, name: &str) -> Option<Value> {
-		let global = self.instance
-			.export_by_name(name)?
-			.as_global()?
-			.get();
+		let global = self.instance.export_by_name(name)?.as_global()?.get();
 
 		Some(global.into())
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use crate::{Error, Value, ReturnValue, HostError, EnvironmentDefinitionBuilder, Instance};
-	use assert_matches::assert_matches;
-
-	fn execute_sandboxed(code: &[u8], args: &[Value]) -> Result<ReturnValue, HostError> {
-		struct State {
-			counter: u32,
-		}
-
-		fn env_assert(_e: &mut State, args: &[Value]) -> Result<ReturnValue, HostError> {
-			if args.len() != 1 {
-				return Err(HostError);
-			}
-			let condition = args[0].as_i32().ok_or_else(|| HostError)?;
-			if condition != 0 {
-				Ok(ReturnValue::Unit)
-			} else {
-				Err(HostError)
-			}
-		}
-		fn env_inc_counter(e: &mut State, args: &[Value]) -> Result<ReturnValue, HostError> {
-			if args.len() != 1 {
-				return Err(HostError);
-			}
-			let inc_by = args[0].as_i32().ok_or_else(|| HostError)?;
-			e.counter += inc_by as u32;
-			Ok(ReturnValue::Value(Value::I32(e.counter as i32)))
-		}
-		/// Function that takes one argument of any type and returns that value.
-		fn env_polymorphic_id(_e: &mut State, args: &[Value]) -> Result<ReturnValue, HostError> {
-			if args.len() != 1 {
-				return Err(HostError);
-			}
-			Ok(ReturnValue::Value(args[0]))
-		}
-
-		let mut state = State { counter: 0 };
-
-		let mut env_builder = EnvironmentDefinitionBuilder::new();
-		env_builder.add_host_func("env", "assert", env_assert);
-		env_builder.add_host_func("env", "inc_counter", env_inc_counter);
-		env_builder.add_host_func("env", "polymorphic_id", env_polymorphic_id);
-
-		let mut instance = Instance::new(code, &env_builder, &mut state)?;
-		let result = instance.invoke("call", args, &mut state);
-
-		result.map_err(|_| HostError)
-	}
-
-	#[test]
-	fn invoke_args() {
-		let code = wat::parse_str(r#"
-		(module
-			(import "env" "assert" (func $assert (param i32)))
-
-			(func (export "call") (param $x i32) (param $y i64)
-				;; assert that $x = 0x12345678
-				(call $assert
-					(i32.eq
-						(get_local $x)
-						(i32.const 0x12345678)
-					)
-				)
-
-				(call $assert
-					(i64.eq
-						(get_local $y)
-						(i64.const 0x1234567887654321)
-					)
-				)
-			)
-		)
-		"#).unwrap();
-
-		let result = execute_sandboxed(
-			&code,
-			&[
-				Value::I32(0x12345678),
-				Value::I64(0x1234567887654321),
-			]
-		);
-		assert!(result.is_ok());
-	}
-
-	#[test]
-	fn return_value() {
-		let code = wat::parse_str(r#"
-		(module
-			(func (export "call") (param $x i32) (result i32)
-				(i32.add
-					(get_local $x)
-					(i32.const 1)
-				)
-			)
-		)
-		"#).unwrap();
-
-		let return_val = execute_sandboxed(
-			&code,
-			&[
-				Value::I32(0x1336),
-			]
-		).unwrap();
-		assert_eq!(return_val, ReturnValue::Value(Value::I32(0x1337)));
-	}
-
-	#[test]
-	fn signatures_dont_matter() {
-		let code = wat::parse_str(r#"
-		(module
-			(import "env" "polymorphic_id" (func $id_i32 (param i32) (result i32)))
-			(import "env" "polymorphic_id" (func $id_i64 (param i64) (result i64)))
-			(import "env" "assert" (func $assert (param i32)))
-
-			(func (export "call")
-				;; assert that we can actually call the "same" function with different
-				;; signatures.
-				(call $assert
-					(i32.eq
-						(call $id_i32
-							(i32.const 0x012345678)
-						)
-						(i32.const 0x012345678)
-					)
-				)
-				(call $assert
-					(i64.eq
-						(call $id_i64
-							(i64.const 0x0123456789abcdef)
-						)
-						(i64.const 0x0123456789abcdef)
-					)
-				)
-			)
-		)
-		"#).unwrap();
-
-		let return_val = execute_sandboxed(&code, &[]).unwrap();
-		assert_eq!(return_val, ReturnValue::Unit);
-	}
-
-	#[test]
-	fn cant_return_unmatching_type() {
-		fn env_returns_i32(_e: &mut (), _args: &[Value]) -> Result<ReturnValue, HostError> {
-			Ok(ReturnValue::Value(Value::I32(42)))
-		}
-
-		let mut env_builder = EnvironmentDefinitionBuilder::new();
-		env_builder.add_host_func("env", "returns_i32", env_returns_i32);
-
-		let code = wat::parse_str(r#"
-		(module
-			;; It's actually returns i32, but imported as if it returned i64
-			(import "env" "returns_i32" (func $returns_i32 (result i64)))
-
-			(func (export "call")
-				(drop
-					(call $returns_i32)
-				)
-			)
-		)
-		"#).unwrap();
-
-		// It succeeds since we are able to import functions with types we want.
-		let mut instance = Instance::new(&code, &env_builder, &mut ()).unwrap();
-
-		// But this fails since we imported a function that returns i32 as if it returned i64.
-		assert_matches!(
-			instance.invoke("call", &[], &mut ()),
-			Err(Error::Execution)
-		);
 	}
 }
